refactor: Convert the Set Iris action to structured implementation.
